refactor(LottoNumber, LottoGame, LottoTicket): move classes to domain package
