+++ src/main/java/lotto/Application.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoTicket;
 import lotto.dto.LottoResult;
 import lotto.dto.LottoWin;
 import lotto.view.InputView;
