+++ src/test/java/LottoNumberTest.java
@@ -1,4 +1,4 @@
-import lotto.LottoNumber;
+import lotto.domain.LottoNumber;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
