+++ src/test/java/LottoVendingMachineTest.java
@@ -1,5 +1,5 @@
-import lotto.LottoNumber;
-import lotto.LottoTicket;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
 import lotto.LottoVendingMachine;
 import lotto.Match;
 import lotto.dto.LottoResult;
